Send xterm escapes directly to terminal
This avoids the problem of sh capturing the escape sequences.
Unfortunately, sending xterm an escape to set modifyOtherKeys=2 is
still glitchy.  It always sends the right escapes, but it randomly
eats the first character sent to the terminal after sending the
escape.

Disabling this feature allows getkey to work reliably under xterm, but
without the precision we'd hope for.