Factor out offset for control characters
For a given ASCII control character, there are as many as eight normal
characters that map to it.  That's because the control key sets the
top three bits of the normal character to 0.  When choosing which of
those normal characters to use for our "Ctrl-..." string, we use an
offset which points into the right ASCII block.  This offset will
vary, so I want it separate.